# docs/openapi.yaml
openapi: "3.0.1"
info:
  title: Face Recognition API
  description: This is the final project for GeekBrains course
  version: 1.0.0
  contact:
    email: no@name.com
...
paths:
  /staff/add:
    post:
      summary: Create new employee
      tags: 
        - Staff
      operationId: CreateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        200:
          description: employee has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEmployeeResponse"
        500:
          description: any error (for mobile contract)
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
  /staff/update:
    put:
      summary: Update employee
      tags: 
        - Staff
      operationId: UpdateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        200:
          description: employee has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateEmployeeResponse"
        500:
          description: any error (for mobile contract)
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Meta: 
      type: object
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    EmployeeBase:
      type: object
      required:
        - name
        - photo_id
        - meta
      properties:
        name:
          type: string
        photo_id:
          type: integer
          format: uint64
          x-go-name: PhotoID
        meta:
          $ref: "#/components/schemas/Meta"
    Employee:
      allOf:
        - $ref: '#/components/schemas/EmployeeBase'
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: uint64
              x-go-name: ID
...
    CreateEmployeeRequest:
      allOf:
        - $ref: '#/components/schemas/EmployeeBase'
    CreateEmployeeResponse:
      allOf:
        - $ref: '#/components/schemas/Employee'
    UpdateEmployeeRequest:
      allOf:
        - $ref: '#/components/schemas/Employee'
    UpdateEmployeeResponse:
      allOf:
        - $ref: '#/components/schemas/Employee'
...